refactor, made code more reusable, exp create
